Removed Floating Pixels on Steampunk Wings (#3046)
* Beta Bugfix

* Removed floating pixels in the Steampunk Wings Images

I am not sure if they were meant to be there. Like if they were meant to be a tiny gear or something.

But since they are only visible in very certain backgrounds I am assuming they were just an accident. So have removed them.

The files got slightly bigger due to the program I saved them in.