Fix conflicting pose issue between the HempRope & FrogtieStrap (#3137)
The pose system seems to expect all pose arrays to be ordered the same,
as it implicitly relies on the last pose being the same no matter which
item allowed it. Change the pose ordering on the HempRope asset so it
matches the one from FrogtieStraps, squashing a bug where the
legs and frogtie would match the kneeling spread pose, but the hemp rope
would display the kneeling one.