[Beta fix] [asset] [Natsuki] Plastic roll update (#3138)
* item code basics

getting back into coding, so trying for a basic item. cant work to the top without getting the basics down

* preview and defining code

item preview added as well as the code needed to name and define the item

* more defining and asset fixes

left out some definitions that were needed to get the item to show, and fixed up some of the alignment with the asset

* include "Type"

missed the TYPE part in the dialog, my bad on that

* issue fixed

didnt quite line up the { and the }.

mental note, fetch origin before merging to avoid fixes within github

* respaced to look neat (hopeful issu fix v2)

* spacing fix on internals

thanks to Jomshir98 for helping me catch a spacing issue, as it was i direct mistake with using github editing. thank you so much for this catch

Co-authored-by: Jomshir98 <jomshir98@protonmail.com>

* Update BondageClub/Assets/Female3DCG/Female3DCG.js

Co-authored-by: Jomshir98 <jomshir98@protonmail.com>

* Size readjustment

makes the wrap a tad smaller

Co-authored-by: Jomshir98 <jomshir98@protonmail.com>