Implement reverse activities
This allows activities to show up when the character you click on wears
something. Adds three new activites for reverse sucking, penetration and buttsex.

The strapon panties block the slot entirely, and I wasn't able to find
a combination of properties, activities or effect that would make it
work realistically.