Cleanup the way activities are defined (#3294)
* Rework how activities are declared

This changes the data so that activities declare their targets instead
of the other way around and orders them depending on their current
dialog ordering, which I used a script to extract.

Activity & AllowActivityOn are gone from AssetGroupDefinition, the
former because of the data change, the latter because it's never used.

There's a new low-level function for getting to the list of defined activities
for a group.

Activity is changed into a single string instead of an array. As far as
I can see, there never were a way to have more than one possible
activity enabled on an asset.

The Target and TargetSelf values are now sorted for simpler visual
checking. Since there's mostly overlap in both target list, allow setting
TargetSelf to true to mean "use the normal target list". This allows removing
the "SelfOnly" prerequisite, since a "SelfOnly" activity can now be written
with a defined TargetSelf list and an empty Target one.

Finally, the SpankingItems required an override for the "Item" activities
so they would show up correctly.

* Change some items from Activity to AllowActivity

Finish cleaning up Activity from Assets and AssetGroups.
This makes it clearer that the activity is supposed to trigger when
used, not when worn (like giving someone one of the handheld items or
applying the shock items). As the rest of the code isn't ready for such
a distinction, rewire the spanking toys to look at those new
AllowedActivities.

* Make the penetration prerequisite more generic

The activity system will now accept Needs-$attribute as a prereq, meaning you
can have activities that only become enabled when an item with a specific
attribute is worn.

* Enables a few missing activities on self

Namely, tickling your own neck, spanking your feet, fisting yourself,
slapping your vulva and clit.

* Reorder the activities

Turns out the automated ordering I had looked like crap in some case, so
I did it manually. It's supposed to match what body part they use, in
order: mouth, hand, feet, whole-body, head, then the item-related ones.

Inside that, they get sorted by softest to harshest, when that makes
sense.