Fix another bug where sounds would not play (#3332)
The IsPlayerInvolved check in ChatRoomMessage had a second part for
handling the case where the player would be at the receiving end of a
message. Since the Action and Activity blocks are mutually-exclusive,
Activity-based sounds wouldn't be triggered in that case.

Refactor that "player involved" code into a helper, and have the audio
system use that to know if a sound should play or not.