Fix UseRemote effect (#3335)
* Hoist condition and reformat slightly

* Hoist common conditions, extend if block, and simplify

* Move "remote disabled per owner rule" to the helper

* Fix an issue with the remote's status

We're not checking the helper's return value, so clicking on the "item"
remote always works.

* Fix issue with UseRemote effect

The UseRemote effect is supposed to change the Use button to a Remote
button. With the code as it is, any extended item (a.k.a "usable") with
that effect would have its Use button shown, preventing the Remote
button from being shown instead.

Fix the issue by checking the item's remote state first, only adding Use
if the item isn't remotely controlled.

* Fix a deadlock issue with UseRemote and the Futuristic Collar

As enabling "Block remotes on the wearer" disables the ability to use
the collar remotely, you get deadlocked out of the collar interface,
and the collar must be removed to be able to fix it.

Remove that effect from the collar, squashing the deadlock.