Fix dereferencing null DOM elements (#3358)
Most callers of ElementValue expect it to return a string, and 
usually run away doing a toUpperCase or some String method call on it. But if the element isn't in the DOM, you get back a null, and
the code crashes.

Log an error but return an empty string, so the code can at
least continue safely.