Fix dereferencing null DOM elements, take 2 (#3376)
* Fix dereferencing null DOM elements

Most callers of ElementValue expect it to return a string, and 
usually run away doing a toUpperCase or some String method call on it. But if the element isn't in the DOM, you get back a null, and
the code crashes.

Log an error but return an empty string, so the code can at
least continue safely.

* Fix issue with non-Chrome browsers

It turns out getAttribute doesn't access the element's internal value,
so this caused the code to return "" instead of the entered value (or
the setAttribute one).