Refactor ChatRoomPublishAction
This changes ChatRoomPublishAction to not be responsible of leaving the
dialog anymore because most callers should know better if they want it
closed or not (there were a couple mismatches with that), and makes the
actual action to publish explicit, so no more guessing what happened
with the two items.

This also makes those "new" actions pass through the struggle minigame,
so you clearly know what to publish when they complete.