Use a PRNG to build the lock pin combination
As we were saving the lockpick combination onto the item, but not
sending the update, a concurrent update from the lockpicked character
(which wouldn't have had the property set as it wouldn't have opened the
lockpicking UI) caused a reset of the lockpicking combination,
apparently randomly.

Fix the bug by building the lockpicking combination using the PRNG, 
seeded with a combination of the lockpicked character's ID, the group
being lockpicked, and the difficulty of the lock, which itself is a
depends on the lockpicker's state (skill and blocking, mostly).

This allows removing the validation handling for that property, and
makes the lock-pick combination unique per lockpicker.