Implement asset file-checking into assetcheck
This is an attempt at making assetcheck filesystem-aware, in that
it will build a tree of all .png files, perform the — dare I say —
hairy transformations needed to turn an asset's layering into a
list of its filenames, and compare that with the tree, reporting any "missing" or "exists" files.

It's still rough around the edges, and there are definitely things
that are not processed correctly (like, unsurprisingly, modular items). But at least it's easy for someone to go over the list of
reported issues and be cleverer.

Known issues:
- modular items as well as some simpler ones like the coffin.
- files that are managed outside of the asset's definition. The kennel's animation files get flagged.
- files that are prefixes of another set, like FuturisticHeels vs. FuturisticHeels2. The former will get the latter's files reported
as exists.