Fix ChatRoomData default value by initializing it to null
This requires a tweak to the ChatRoomSync processing, as we were
processing the synced rooms' characters before setting the actual room
data, causing a crash in CharacterLoadOnline when it tried to find the
character, as there would have been no room (and thus no characters to 
search) yet.

Note that I'm completely confused as to how the code manages to work,
since it means that function had been checking against a potentially
stale version of the room the player was previously in, instead of the
current room they're joining.