Centralize account-checking regexes and simplify the code
This moves the few regexes used for checking account & character name,
password, email and reset number into Server.js (since it's easier to
check that the server ones match). It also removes the length checks in
favor of doing that in the regex itself.

Note that the email regex wasn't consistent between account creation and
password reset (the latter was rejecting stuff like + or -). The check
itself is moved to a helper so we can do basic format checks (has a @
and a . in the 2nd part).