Split out metadata and substitution processing
We previously had two different blocks responsible for this, one for
actions/server messages, and one for activities, and those behaved
differently.

Consolidate all that into two helper function, one that looks at
ChatMessage.Dictionary and extracts any relevant information and
requested text substitutions, and one that applies the given list of
substitutions to a message.