Fix the Vacbeds triggering stimulation events
Having the IsPlugged effect results in it being taken into account when
applying stimulation events, leading to a non-sensical "The clear vacbed
inside your body feels awkward as you move.".

Remove the effect, so that doesn't happen anymore, and add a new
ButtChaste effect to indicate that it's not accessible but still empty.

Also whitelist the vulva & clit slots, for consistency.