Rework the two pet suits so they follows the actual character pose
Since AllFours now appears in the pose menu for those, the type option
to force a pose change became unnecessary, and was causing a visual bug
where the strap wouldn't show in the AllFours pose (since the type
wasn't actually set). This fixes both suits so they work as they were
previously but using actual character poses.

The only caveat that the pose change notification is gone, because I
feel that should be done when the user clicks the pose buttons.