Remove allowed-check when updating the crafting dummy
That check was made to limit visual oddities from getting drawn by
checking the crafted item's prerequisites before wearing it, a good
example was custom blankets, which would require a bed first.

As it turns out, if you're clothed in any way, this causes simpler
assets like the anal hook to be skipped (as panties are in the way),
resulting in a crash later when trying to pick colors on an item not
actually worn.

Fix the issue by removing the check, allowing any asset to be worn, even
if it's visually non-sensical.