Fix a bug with the pose menu not reflecting the available poses
Previously, the pose menu was generated the first time you'd open/draw
the menu, and nothing would invalidate it. As the AllFours pose might
or might not be there, this meant that opening the menu while not in
a pet suit would cause the option to never show up — or the reverse,
the option would be there without a petsuit being worn.

Fix the issue by adding a Load callback to the menu system that's
responsible for initial setup, and use it to rebuild the pose menu when
you switch to that option.