Fix the broken arousal system because of a missing group
As we weren't looking at the correct metadata entry, the arousal system
would be handed an undefined value, leading to the activity cap being
applied even on area where orgasms should happen.

Fix the issue by passing in the correct metadata entry. As dynamic items
uses another key, also pass that one in if the first one is missing.