HairBack Fixes by Kyra
My first commit!

This is a change to many existing HairBack assets for compatibility with the new smaller Head asset and the shorter HairFront assets that accompany it.
This is currently done for the regular pose, AllFours and Hogtied.
Suspension is still to be done.
They will likely still not look aesthetically pleasing in combination with short HairFront assets, but any awkward gaps between the head and hair should now be filled.