HairBack Fixes for Suspension Pose by Kyra
This is a change to many existing HairBack assets for compatibility with the new smaller Head asset and the shorter HairFront assets that accompany it.
This change does this for the Suspension pose.
This is already done for the regular pose, AllFours and Hogtied.
They will likely still not look aesthetically pleasing in combination with short HairFront assets, but any awkward gaps between the head and hair should now be filled.