Gag Strap Extensions by Kyra
A slight adjustment to the length of certain gag straps in order to fit them closely to the shape of the new Head asset, so that they show correctly combined with short hairstyles.
This covers assets where permissions had been granted during an earlier pass at short hairstyles.
Additionally, added a 'tight' option for the Harness Ballgag XL