Fix a crashing bug with the vibes
As that line replaces one of the AllowType entries, this makes the guard
in VibratorModeLoad fail to work, as by default a vibe is in the Off
mode, triggering a character refresh storm and either a server kick or a
stack overflow error.

Since Off is the real mode value there, remove that line and update the
few vibrating items AllowType properties that were using the other
value.