Hair Accessory Refits by Kyra
This change expands Hair Accessory items to work with short hairstyles.
This is achieved by marking short hair with a 'ShortHair' attribute and having particular hair accessory assets react to this using Ellie's ShowForAttribute and HideForAttribute functionality.
Existing assets have been renamed to add a _Long layer suffix where required. Similarly _Short has been used for the corresponding layer assets for short hair.
In many cases, the _Short layer is a duplicate of _Long which is positioned differently. In some cases, the _Short layer asset has been resized in order to fit the head better.
As a result of this change, all Hair Accessory items should now properly fit the head when short hairstyles are used.