Custom Latex Hood Cover Addition
Adds an optional additional piece to the custom latex hood.
This is a second hood or 'cover' which zips over the first.
It is separately colourable, and has transparent options.
It hides a few different item slots, and also forces the front hair to be hidden, all due to visual issues when these are left visible.