Fix a bug impacting an activity's arousal and sfx
As part of the larger cleanup I made to activity-handling in the
handheld rework, `Asset.DynamicActivity()` was made almost useless.
Right now, it only acts as getter for `Asset.Activity`, and is only
useful as part of triggering an arousal effect when applying some items
(those that have an explicit `Activity` set), for example the anal beads.

As there aren't anymore dialog keys for it (old shock system), and the
rework from Ellie introducing an issue where it would overwrite the true
activity name with `undefined` using that method, remove the rest of the
leftovers so that at least it doesn't cause a failure down the line.

This fixes item activities not working correctly, or missing sound
effects on the handhelds, as they need the `ActivityName` in metadata to
work, and for example the crop doesn't cause an activity when pulling it
out (what `Asset.Activity` is about).