Cropped Hoodie by Kyra
Adds a hoodie for the cloth slot.
This is currently named the 'Cropped Hoodie' in the inventory and has options for both hood shape+position and for length of the garment.
Currently only 'cropped' and 'bolero' lengths exist, but a longer option including a working front pouch is planned.

The hood uses alpha masks to create the impression of hair and head items being inside the hood.