Refactor the "ChatRoomLeave" calls into a dedicated function
That function is now responsible for resetting *all* the various
room-related state, like leashing, lent lockpicks, last chatroom visited
as well as clearing the UI.

As some of the previous calls were also clearing the cached online
characters, most of those were revisited so it would only happen for
"long-term" options (calling the maids for help, prison/GGTS isolation).
All the other callsites might not actually need that to happen and
benefit from not having to recache characters.