Short Hair Depth Layers
Adds a 'front' layer to most 'short' hairstyles where due to the smaller space, many hair accessories seemed to crowd the head and obscure too much of the hair.
This 'front' layer is drawn just above hair accessories and should hopefully give the appearance of depth. Due to the shapes used, many are not really suitable for additional colouring options - additional layer assets should be made if colouring options are expanded for these hairstyles.