Add back the pulled leashes
This restores the held leashes removed in !3825, by adding their
slightly tweaked assets to the normal ones and using dynamic drawing to
swap them in when they get grabbed.

As part of knowing if a leash is held or not, a new `IsLeashed` effect
was added, a bit of cleanup splitting out the leash "RPC" into
dedicated functions, and a new one at the character level to handle
refreshing the leash state. This also fixes a bug, breaking the leash
link in case we aren't wearing a leash anymore.