Fix a crash in the Cafe
The intensity setter function was deleted in 57d4c4d, hence the game
would crash on attempting to increase it.

Rename one of VibratorMode.js function for consistency, and provide an
easy-to-use setter that takes a VibratorMode and updates the item's
state appropriately. Note that it's a slight hack, as the Inflatable
Dildo doesn't have the correct archetype anyway, so we also update the
Type to be correct.