Separate Left and Right Sock Options
This adds groups for the Left and Right legs, including most currently available socks and stockings
Uses buygroups so that the 'same' set of socks do not have to be bought multiple times.
Independent of the existing "Socks" group for the time being, though it may make sense to distribute "Socks" differently at some future point.