Reduce Priority for the Cropped Hoodie's Front Layers
There was a small clash between the front layers of the hoodie (using priority 60) and the front layers of the pet bed (which uses the group default priority 56) causing those parts of the hoodie to layer on top of the pet bed.
Sending the hoodie's hood to 56 had unwanted consequences of allowing some headwear to draw on top.
The layering as is was too cramped, so moving the front piece of the pet bed to 58 and the hoodie's hood to 57