Fix another problem with locks and option switching
This is a follow-up to !3874, as that change actually caused the Love
Chastity Belt (an owner-only item) to break, as it would see a lock and
prevent any sort of change since its ChangeWhenLocked property would
actually be `undefined` (while it should default to true).
The other change makes the check more sturdy against invalid data by
making sure the current option is actually lockable.