Revert "Merge ParentColor with InheritColor"
This reverts the following commits:
- 5ebecb8f9b7519111833208116b6014bcd43ab0d
- d03a1375a167e988fd2ee759ff565ffe6a69569f
- 02b73751c1f5b5a8cf8220b7d15245e96393201a

It was determined during beta that the color inheriting would also 
apply between the Cloth/Cloth lower groups, leading to a color change
from the former changing colors from the latter. Hence the revert, as
I'm not sure it's possible to have it both ways.