Fix a couple issues where the leash could break
As `ChatRoomCanBeLeashedBy` wasn't checked yet, this would allow someone
to send out a leash message, triggering a break-up even though they
should not actually be able to leash us. Another one is that we need to
refresh our leash state if the check fails, which can happen when
entering a leash-blocked room, or the current one gets changed.

Fix the issues by only breaking the leash when we're about to swap
leashers and refresh the leash unconditionally, and refreshing the leash
state when the room properties change.