Fix lover's lock removal logic
This fixes an issue where the login validation code would make use of
`Property.MemberNumber`, but nothing in the project actually set that 
value, leading to lover items never getting removed (as far as that
function is concerned).

As it turns out, the client-side validation ran just before that, and
has actually taken care of yanking the locks themselves. Hence, remove
that part and just keep the love leather collar handling.