Fix another issue with ChangeWhenLocked
As reported on Discord, the OTN plug gag gets blocked if you lock it, 
which is a bug. This removes the duplicated check from one of the places
it's being done, removes the temporary data variable for it, and corrects
the check to account for cases where it's not specified.