Use a callback to handle post-struggle dialog updates
This adds a new callback that gets set when the minigame starts and is called
when either a success or an interruption happens. This allows the result of the
minigame completing (or being interrupted) to be moved back into Dialog itself,
keeping more neatly separated. The only exceptions to this right now are
lent lockpicks and the expression changes.