Use a new global variable to track the dialog mode
This changes the dialog to use one variable instead of one for each of locking,
looking at a lock, permissions, activities, crafting, coloring, extended
screens, the new tightening screen, selecting a struggle minigame and
struggling, making it much easier to keep mode checks consistent.

Note that the old variables are kept around, but marked as deprecated as they
are now useless. This is mostly done to stop crashes in extensions or scripts
if they try to access them.