Separate Gag Kissmark as its own Item
Separates the gag kissmark as its own item.
Available only to gagslot 3
Wearing a covered gag is a requirement (new 'blockedmouth' prerequisite added to permit this)
Will be unequipped when a gag is unequipped

Other kissmarks available as clothing in both the cloth accessory and mask slots.

All bundled under 'smoochy lipstick' in the shop, which is available with the other cloth accessory items.