Split the minigame's difficulty into two variables
Since the old MiniGameDifficulty gets used both as a string or as a
number depending on the minigame, TS screams any type a use doesn't
match what the game actually expects. Fix the issue by checking the type
of the value coming in, store it in a different variable for each case,
and search/replace move all the string-users to the new one.