TS: Change C.FocusGroup to be an AssetItemGroup
As this is supposed to mean "the group currently selected", and the UI
only exposes item groups, change the property so the type matches.

The only two exceptions found where Appearance.js, because it reuses
Dialog's inventory view for clothes, and the Shop, which tracks the
selected section using that group.