Implement New OverridePriority on CustomLatexHood
Changes the Custom Latex Hood to use Estsanatlehi's new OverridePriority functionality!
This allows the hood cover to remain at priority 50 when the rest of the item changes, which resolves a visual bug between the cover and blindfolds.
This has allowed for the hood cover to no longer be mutually exclusive with blindfolds