Fix a bug with the DictBuilder method crashing in conditional blocks
As the check handling conditional dicts only concerned _addEntry,
trying to protect say targetCharacter from seeing a null would still
cause a crash when it tried to get the member number.

Fix the issue by hoisting _condition up with a default to `true` and use
it as a guard in every method so that a falsey condition does result in
completely skipping over the method's behavior.