Improve the login screen
This shuffles buttons around so that the language selector isn't at the
"default validation" spot anymore, instead pushing it to the bottom of
the screen, changes it to show language names instead of the translation
for language (which is not a word that's often known when facing a
mismatched UI) and adds a flag icon to make it generally obvious it's
country-related. The login button gets moved to the now freed 
"default validation" spot, and "reset password" gets to have the spot
for "cancel", since that feels more related than anything.