Avoid calls to CraftingUpdatePreview()
Calling CraftingUpdatePreview() invokes an O(2n ** 2) loop againat the
entire Asset list, which causes text entry in the Name and Description
fields to be incredibly sluggish and frustrating.

The correct solution would be to refactor CraftingAppliesToItem() and
everything that calls it (including CraftingUpdatePreview()) to not
require scanning the entire Asset list a zillion times.  Until then:

Alter CraftingKeyUp() to call CraftingUpdatePreview() only when
something in the preview imagery actually changes.