Fix missing vibe sound effects
This updates the audio helper function to use metadata instead of string
wrangling to detect which asset vibration mode was changed, and
updates the list to use actual asset names instead of whatever it was
before.

Note that items that use multi-module vibrators are still iffy: they'll
play a sound, but it might not be the correct one since there's no way
right now to tell which module caused the vibe change.