Remove leftover nipple items after Asylum treatment.
If, as a Nurse/Doctor, you perform an orgasm therapy, items are added to
the patient's nipples but not properly removed after therapy has
concluded.  The items then appear on all subsequent patients after
they've stripped, regardless of the therapy they're undergoing.

Remove any items on the patient's nipples at the conclusion of the
therapy session.