Fix a bug that would cause GGTS to strike on pose changes
As we were relying on the JSON serialization order of Player.Pose to be
consistent (which it is not, since it depends on the order of the poses
clicked in the UI), doing a JSON-string compare of the current value
with the saved one would occasionally fail for no obvious reason.

Fix the issue by comparing pose arrays directly for unordered
equality, which sidesteps the serialization problem.