Fix a crash with the advanced vibe modes
As we're trying to draw icons, vibe options have a special structure
that doesn't really match what normal typed items do. Make sure to
match them by name instead of a non-existent `Property.Mode` property.

This skips a crash later in ExtendedItemRequirementCheckMessageMemo
when it tries to get the option name when checking prerequisites.