Fix another bug with advanced modes
When setting the extended option for an advanced mode, we would reset 
Mode to null/undefined, leading to the state getting lost (buttons not
selecting, and vibe audio not playing).

Ensure we pass the actual mode when setting the extended option so that
the tracking done by `VibratorModeScriptDraw` works correctly.