Follow chains of CopyConfig when loading
It turns out there's at least a chain of two:
HighStyleSteelCuffs > SteelCuffs > LeatherCuffs.
Since BC ignores the second CopyConfig, some of the configuration
ends up with default values, and you get a missing translation for the
header afterward.

Fix the issue by doing a loop over the chains' "links" through `CopyConfig`, tracking assets' group/name as we go so we don't end up in a cycle.