Make AssetCheck always perform dialog checks
Since having a CopyConfig doesn't mean values haven't been overridden,
we have to look at the dialog keys.

Also as a sanity measure, we overwrite the loaded config with the fully
resolved one to save on code.

Lastly, the previous dialog checker is removed, since the interface
changed and it's now useless.