Make the game's main loop self-contained
This moves the "recursive" requestAnimationFrame call, as well as the
code responsible for making the global timer "tick" back into the
entry point, for clarity.

This also fixes an oddity with how `CurrentTime` was calculated, as the
math would make it a large negative number until the server time was
received and everything caught up.