Make the Padded & Paw Mittens into typed assets
This removes the MittensChain1 "virtual" item, as well as the script
files for those two in favor of using the extended asset system
directly.

Note that most of the duplicated, identical layers for those were also
deleted, since the upper body size has no impact on hands' size.

This also adds the "chainable" attribute to whatever asset has something
that looks like an attachment point for the chain (the harness bra,
leather strap harness & studded harness).