Fix a bug where overflowing icons couldn't be clicked
This makes the menu click handler return whether a button was triggered,
and moves the check from the main handler around so it happens before
handling clicks outside the dialog area (unselecting the focused group),
making it possible to click on icons that overflow the item area.

Since some screens forgo the buttons in DialogMenuButton entirely,
special-case those so that the handler doesn't prevent them from doing
their own thing. Lastly, fix a bug with the ActivityMode Exit button
returning straight to the dialog and not the inventory view.