Split out the single color mode into three separate ones
As there are two color pickers in use, one that requires an item
(ItemColor), and one that's just an UI for color selection, having
them have their specific type makes more sense. And on top of that,
since expressions can be colored, and need to restore the group they
started from, split that mode out as well.