Refactor the modular type generation to happen faster
We're taking 6+ seconds at startup to load the assets, the slowdown
being caused by the generation of their layers' `AllowType` from their
`AllowModuleTypes` configuration. As it turns out, the code suffers from
a bunch of issues: it regenerates the whole `AllowType` property for
each layer, which is the cartesian product of each module * each option,
and some of theses are *massive* already (the Custom Latex Hood is
upward of 30000 types).

We can be quicker than this, and this removes the whole generation step
for the layers, instead relying on the layer visibility check to decide
when to see which ones. This takes the startup time down from 6+s down
to 0.5s.