Implement PortalLink support between the panties and the tablet
This adds the necessary features to be able to remotely do things with
PortalLink-enabled transmitters and receivers, like the Portal Panties
and Tablet. It is set up using a shared code between both "peripherals",
and the tablet remotely reads the list of available functions supported by
its target.

Supported features are remote locking, chastity cycling, and some actions
through the portal when enabled. All of that is managed through a set of
attributes which the transmitter queries once it found a code match.

There's a new hidden lock added that is used to enforce usage of the
tablet. It is marked as exclusive, so the locker can remove it manually, but other players will have to go through the tablet.

The panties were also re-layered to make a bit more sense visually.