Allow Colorization of Highlights layers on the Custom Latex Hood
This has been on the to-do list for a while and was initially planned to bundle with options for different shine finishes.
The assets for different shine finishes never really materialised (though I'll add them if they do!) so it's about time to add this small feature as a QoL addition.

Adds colorization to the highlights layers of the custom latex hood.
This has been done using a colour group of 1 on the back end, for the sake of simplicity if in future there's a need to expand this to multiple types or areas of highlight.