Provide helper function to show, hide, and append to the chat log
This consolidates all the code doing element display changes into one
place, instead of being copy-pasted all over.

The only functional change is that ChatRoomRun now forcibly restores
the chatroom UI when the elements are hidden, and that the creation
function sets up the input field position as well (the rects having
been moved to a global, since there were two different rects used in).

This is definitely a stopgap measure, as the expectation would be for
whichever screen that hid those elements to restore them on unload (but
almost none make use of that). At least it prevents the UI from
disappearing entirely.