Prevent a rate-limit crash when clicking around in permission mode
As CharacterRefreshDialog calls DialogChangeMode to refresh it in
permission mode when receiving character sync data, trying to send an
update as part of it causes a recursive loop, crashing the game and
slowing down the other participants because of the message storm.

As it turns out, item permissions are updated and broadcast as soon as
an item is clicked, so that whole block is only useful in breaking
the game.