Fix a crash with the training belt's deny mode
As its script draw hooks yanks us out of the dialog while the character
is being drawn, this causes CurrentCharacter to become null, leading to
a crash as we're still getting through the dialog draw function.

Hence, add another null check after the characters are drawn, skipping
the rest of the function if that happens.